src/multi-ai-agent-setup/requirements.txt
azure-functions
azure-functions-durable
openai
azure-cosmos